+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.0.0] - 2018-06-12
+### Changed
+- Release version v1
 
 ## [0.2.0] - 2018-06-04
 ### Added
-Add enhaced class with better formated data.
-
+- Add enhaced class with better formated data.
 
 ## [0.1.0] - 2018-06-01
-Project start
+- Project start